@@ -34,7 +34,7 @@ class BoardTest extends TestCase
     {
         parent::setUp();
 
-        $this->user = factory(User::class)->create();
+        $this->user = User::factory()->create();
 
         $this->be($this->user);
 
