@@ -2,12 +2,12 @@
 
 namespace App\Http\Controllers;
 
-use App\Board;
-use App\Card;
-use App\Http\Requests\TaskRequest;
+use App\Models\Board;
+use App\Models\Card;
+use App\Http\Requests\Task\TaskRequest;
 use App\Repositories\Task\TaskRepositoryInterface;
-use App\Task;
-use Lang;
+use App\Models\Task;
+use Illuminate\Support\Facades\Lang;
 
 class TaskController extends Controller
 {
@@ -29,8 +29,7 @@ class TaskController extends Controller
      * Create a new controller instance.
      * Only auth users can see.
      *
-     * @param  \App\Repositories\Task\TaskRepositoryInterface  $taskRepository
-     * @return void
+     * @param TaskRepositoryInterface $taskRepository
      */
     public function __construct(TaskRepositoryInterface $taskRepository)
     {
@@ -41,7 +40,9 @@ class TaskController extends Controller
     /**
      * Show the form for creating a new resource.
      *
-     * @return \Illuminate\Http\Response
+     * @param Board $board
+     * @param Card $card
+     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View
      */
     public function create(Board $board, Card $card)
     {
@@ -51,8 +52,10 @@ class TaskController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \App\Http\Requests\TaskRequest  $request
-     * @return \Illuminate\Http\Response
+     * @param TaskRequest $request
+     * @param Board $board
+     * @param Card $card
+     * @return \Illuminate\Http\RedirectResponse
      */
     public function store(TaskRequest $request, Board $board, Card $card)
     {
@@ -71,8 +74,8 @@ class TaskController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  \App\Task  $task
-     * @return \Illuminate\Http\Response
+     * @param Task $task
+     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View
      */
     public function show(Task $task)
     {
@@ -84,8 +87,8 @@ class TaskController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  \App\Task  $task
-     * @return \Illuminate\Http\Response
+     * @param Task $task
+     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View
      */
     public function edit(Task $task)
     {
@@ -97,9 +100,9 @@ class TaskController extends Controller
     /**
      * Update the specified resource in storage.
      *
-     * @param  \App\Http\Requests\TaskRequest  $request;
-     * @param  \App\Task  $task
-     * @return \Illuminate\Http\Response
+     * @param TaskRequest $request
+     * @param Task $task
+     * @return \Illuminate\Http\RedirectResponse
      */
     public function update(TaskRequest $request, Task $task)
     {
@@ -113,10 +116,8 @@ class TaskController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Task  $task
-     * @return \Illuminate\Http\Response
-     *
-     * @throws \Exception
+     * @param Task $task
+     * @return \Illuminate\Http\RedirectResponse
      */
     public function destroy(Task $task)
     {
