@@ -2,7 +2,7 @@
 
 namespace App\Listeners;
 
-use App\Card;
+use App\Models\Card;
 use App\Events\BoardCreateEvent;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Queue\InteractsWithQueue;
